bot/requirements.txt
python-telegram-bot
python-docx
python-dotenv
